client/requirements.txt
requests
python-jose[cryptography]
cryptography
pyinstaller
